pyproject.toml
[tool.kedro]
package_name = "dl4thermo"
project_name = "DL4Thermo"
project_version = "0.18.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
known_third_party = "kedro"

[tool.pytest.ini_options]
addopts = """
--cov-report term-missing \
--cov src/dl4thermo -ra"""

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]


[tool.poetry]
name = "dl4thermo"
version = "0.1.0"
description = ""
authors = [
    "Kobi Felton <kcmf2@cam.ac.uk>",
    "Jan Ritting <jan.rittig@rwth-aachen.de>",
    "Lukas Rasspe <lukas.rasspe@ltt.rwth-aachen.de>",
]
packages = [{ include = "dl4thermo", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.8,!=3.9.7,<3.11"
matplotlib = "^3.5.1"
kedro = "^0.18.2"
jupyterlab = "^3.3.2"
wandb = "^0.12.16"
rdkit-pypi = "^2022.3.2"
molmass = "^2021.6.18"
Pint = "^0.19.2"
pydantic = "^1.9.1"
tqdm = "^4.64.0"
aiohttp = "^3.8.1"
lxml = "^4.9.0"
python-dotenv = "^0.20.0"
seaborn = "^0.11.2"
chemprop = "^1.5.1"
tensorboard = "^2.9.1"
fastparquet = "^0.8.1"
llvmlite = "^0.38.1"
umap-learn = "^0.5.3"
pystow = { version = "^0.4.5", optional = true }
matscipy = { version = "^0.8.0rc10", allow-prereleases = true }
schnetpack = "^2.0.0"
molplotly = "^1.1.5"
joblib = "^1.2.0"
latextable = "^1.0.0"
feos = "^0.4.1"
salib = "^1.4.7"
ray = { version = "^2.2.0", extras = ["default"] }
lmfit = "^1.1.0"
pura = "^0.2.4"
optuna = "^3.1.0"
pytorch-lightning = ">1.8.0,<2.0.0"
svgutils = "^0.3.4"
cairosvg = "^2.5.0"

[tool.poetry.extras]
stout = ["unicodedata2", "pystow"]

[tool.poe.tasks]
install-torch-cpu = "pip install torch>=1.13.0 -f https://download.pytorch.org/whl/torch_stable.html"
install-pyg-cpu = "pip install torch-scatter torch-sparse torch-cluster torch-spline-conv torch-geometric"
install-pyg-cpu-git = "pip install git+https://github.com/pyg-team/pyg-lib.git git+https://github.com/rusty1s/pytorch_scatter.git git+https://github.com/rusty1s/pytorch_sparse.git git+https://github.com/rusty1s/pytorch_cluster.git git+https://github.com/rusty1s/pytorch_spline_conv.git"
install-pyg = ["install-torch-cpu", "install-pyg-cpu", "install-pyg-cpu-git"]


[tool.poetry.group.dev.dependencies]
poethepoet = "^0.16.5"
ipykernel = "^6.9.1"
isort = "^4.3.21"
autoflake = "^1.3.1"
black = "^22.3.0"
mypy = "^0.761"
vulture = "^1.2"
pytest = "^6.0.0"
pytest-cov = "^2.8.1"
pytest-xdist = "^1.31.0"
kedro-viz = "^4.4.0"
nbdime = "^3.1.1"


[tool.poetry.group.cosmo_rs.dependencies]
cosmopy = { git = "https://github.com/sustainable-processes/COSMOpy.git" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
